Update examples to longs, still need to investigate error files.

+++ trunk/examples/datasets/H5Ex_D_Gzip.java
@@ -82,10 +82,10 @@ public class H5Ex_D_Gzip {
     }
 
     private static void writeGzip() {
-        int file_id = -1;
-        int filespace_id = -1;
-        int dataset_id = -1;
-        int dcpl_id = -1;
+        long file_id = -1;
+        long filespace_id = -1;
+        long dataset_id = -1;
+        long dcpl_id = -1;
         long[] dims = { DIM_X, DIM_Y };
         long[] chunk_dims = { CHUNK_X, CHUNK_Y };
         int[][] dset_data = new int[DIM_X][DIM_Y];
@@ -185,9 +185,9 @@ public class H5Ex_D_Gzip {
     }
 
     private static void readGzip() {
-        int file_id = -1;
-        int dataset_id = -1;
-        int dcpl_id = -1;
+        long file_id = -1;
+        long dataset_id = -1;
+        long dcpl_id = -1;
         int[][] dset_data = new int[DIM_X][DIM_Y];
 
         // Open an existing file.
