+++ trunk/examples/datasets/H5Ex_D_Nbit.java
@@ -83,11 +83,11 @@ public class H5Ex_D_Nbit {
     }
     
     private static void writeData() throws Exception {
-        int file_id = -1;
-        int filespace_id = -1;
-        int dataset_id = -1;
-        int dtype_id = -1;
-        int dcpl_id = -1;
+        long file_id = -1;
+        long filespace_id = -1;
+        long dataset_id = -1;
+        long dtype_id = -1;
+        long dcpl_id = -1;
         long[] dims = { DIM_X, DIM_Y };
         long[] chunk_dims = { CHUNK_X, CHUNK_Y };
         int[][] dset_data = new int[DIM_X][DIM_Y]; 
@@ -144,9 +144,9 @@ public class H5Ex_D_Nbit {
     }
 
     private static void readData() throws Exception {
-        int file_id = -1;
-        int dataset_id = -1;
-        int dcpl_id = -1;
+        long file_id = -1;
+        long dataset_id = -1;
+        long dcpl_id = -1;
         int[][] dset_data = new int[DIM_X][DIM_Y]; 
 
         // Open an existing file.
